+++ automation/src/main/java/org/greenplum/pxf/automation/testplugin/HBaseAccessorWithFilter.java
@@ -3,10 +3,13 @@ package org.greenplum.pxf.automation.testplugin;
 
 import org.apache.hadoop.hbase.HBaseConfiguration;
 import org.apache.hadoop.hbase.HConstants;
-import org.apache.hadoop.hbase.client.HTable;
+import org.apache.hadoop.hbase.TableName;
+import org.apache.hadoop.hbase.client.Connection;
+import org.apache.hadoop.hbase.client.ConnectionFactory;
 import org.apache.hadoop.hbase.client.Result;
 import org.apache.hadoop.hbase.client.ResultScanner;
 import org.apache.hadoop.hbase.client.Scan;
+import org.apache.hadoop.hbase.client.Table;
 import org.apache.hadoop.hbase.filter.Filter;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.greenplum.pxf.api.OneRow;
@@ -55,7 +58,8 @@ public class HBaseAccessorWithFilter extends BasePlugin implements Accessor {
     private static final TreeTraverser TRAVERSER = new TreeTraverser();
 
     private HBaseTupleDescription tupleDescription;
-    private HTable table;
+    private Connection connection;
+    private Table table;
     private List<SplitBoundary> splits;
     private Scan scanDetails;
     private ResultScanner currentScanner;
@@ -109,6 +113,9 @@ public class HBaseAccessorWithFilter extends BasePlugin implements Accessor {
     @Override
     public void closeForRead() throws Exception {
         table.close();
+        if (connection != null) {
+            connection.close();
+        }
     }
 
     /**
@@ -157,7 +164,8 @@ public class HBaseAccessorWithFilter extends BasePlugin implements Accessor {
     }
 
     private void openTable() throws IOException {
-        table = new HTable(HBaseConfiguration.create(configuration), context.getDataSource().getBytes());
+        connection = ConnectionFactory.createConnection(HBaseConfiguration.create(configuration));
+        table = connection.getTable(TableName.valueOf(context.getDataSource()));
     }
 
     /*
@@ -203,7 +211,7 @@ public class HBaseAccessorWithFilter extends BasePlugin implements Accessor {
     private void createScanner() throws Exception {
         scanDetails = new Scan();
         // Return only one version (latest)
-        scanDetails.setMaxVersions(1);
+        scanDetails.readVersions(1);
 
         addColumns();
         addFilters();
@@ -219,8 +227,8 @@ public class HBaseAccessorWithFilter extends BasePlugin implements Accessor {
         }
 
         SplitBoundary region = splits.get(currentRegionIndex);
-        scanDetails.setStartRow(region.startKey());
-        scanDetails.setStopRow(region.endKey());
+        scanDetails.withStartRow(region.startKey());
+        scanDetails.withStopRow(region.endKey());
 
         currentScanner = table.getScanner(scanDetails);
         return true;