+++ server/pxf-hbase/src/main/java/org/greenplum/pxf/plugins/hbase/HBaseDataFragmenter.java
@@ -20,15 +20,13 @@ package org.greenplum.pxf.plugins.hbase;
  */
 
 import org.apache.hadoop.hbase.HBaseConfiguration;
-import org.apache.hadoop.hbase.HRegionInfo;
 import org.apache.hadoop.hbase.HRegionLocation;
-import org.apache.hadoop.hbase.ServerName;
 import org.apache.hadoop.hbase.TableName;
 import org.apache.hadoop.hbase.TableNotFoundException;
 import org.apache.hadoop.hbase.client.Admin;
 import org.apache.hadoop.hbase.client.Connection;
 import org.apache.hadoop.hbase.client.ConnectionFactory;
-import org.apache.hadoop.hbase.client.HBaseAdmin;
+import org.apache.hadoop.hbase.client.RegionInfo;
 import org.apache.hadoop.hbase.client.RegionLocator;
 import org.greenplum.pxf.api.model.BaseFragmenter;
 import org.greenplum.pxf.api.model.Fragment;
@@ -80,8 +78,6 @@ public class HBaseDataFragmenter extends BaseFragmenter {
     @Override
     public List<Fragment> getFragments() throws Exception {
 
-        // check that Zookeeper and HBase master are available
-        HBaseAdmin.checkHBaseAvailable(configuration);
         connection = ConnectionFactory.createConnection(configuration);
         Admin hbaseAdmin = connection.getAdmin();
         if (!HBaseUtilities.isTableAvailable(hbaseAdmin, context.getDataSource())) {
@@ -123,7 +119,7 @@ public class HBaseDataFragmenter extends BaseFragmenter {
     }
 
     private void addFragment(HRegionLocation location, Map<String, byte[]> userData) throws IOException {
-        HRegionInfo region = location.getRegionInfo();
+        RegionInfo region = location.getRegion();
         HBaseFragmentMetadata metadata = new HBaseFragmentMetadata(region, userData);
         Fragment fragment = new Fragment(context.getDataSource(), metadata);
         fragments.add(fragment);